requirements.txt
numpy==1.23.5
matplotlib==3.7.1
streamlit==1.23.1
pillow==9.5.0
tqdm==4.65.0